Renamed files for `JoltGeneric6DOFJointImpl3D`
